move sign pkg to pkg/sign
