+++ cmd/cosign/cli/sign.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/sigstore/cosign/cmd/cosign/cli/generate"
 	"github.com/sigstore/cosign/cmd/cosign/cli/options"
-	"github.com/sigstore/cosign/cmd/cosign/cli/sign"
+	"github.com/sigstore/cosign/pkg/sign"
 )
 
 func Sign() *cobra.Command {
