+++ cmd/cosign/cli/signblob.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/sigstore/cosign/cmd/cosign/cli/generate"
 	"github.com/sigstore/cosign/cmd/cosign/cli/options"
-	"github.com/sigstore/cosign/cmd/cosign/cli/sign"
+	"github.com/sigstore/cosign/pkg/sign"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
