+++ pkg/sign/sign_blob.go
